Changed migration file to bigint.

+++ db/migrate/20190611095700_create_occupations.rb
@@ -2,7 +2,7 @@ class CreateOccupations < ActiveRecord::Migration[5.2]
   def change
     create_table     :occupations , :primary_key => :occupation_id do |t|
     	t.bigint      :person_id, null: false
-    	t.integer      :occupation, null: false
+    	t.bigint      :occupation, null: false
       t.bigint        :creator
     	t.boolean      :voided, null: false, default: 0
     	t.integer      :voided_by
